+++ backend-src/testDocuments.ts
@@ -1,15 +1,8 @@
 import { BaseStoreDataType } from "../src/CoreDataLake";
 import { virtualRootId } from "../src/Workflowish/mvc/model";
-export const testDocuments: Record<string, ()=>BaseStoreDataType > = {
+export const testDocuments: Record<string, () => BaseStoreDataType> = {
     "over4096Items": (): BaseStoreDataType => {
-        const items: BaseStoreDataType = {
-            [virtualRootId]: {
-                lastModifiedUnixMillis: 0,
-                data: "",
-                children: [],
-                collapsed: false
-            }
-        };
+        const items = getDataBlockWithVirtualRoot();
         const createStack = Array(8).fill(0).map((val, idx) => [val, idx, virtualRootId]) as [number, number, string][];
         while (createStack.length > 0) {
             const top = createStack.pop() as [number, number, string];
@@ -28,5 +21,38 @@ export const testDocuments: Record<string, ()=>BaseStoreDataType > = {
             }
         }
         return items;
+    },
+    "ohNoADuplicate": (): BaseStoreDataType => {
+        const items = getDataBlockWithVirtualRoot();
+        items["test"] = {
+            lastModifiedUnixMillis: 0,
+            data: "Hello world",
+            children: ["test3"],
+            collapsed: true
+        }
+        items["test2"] = {
+            lastModifiedUnixMillis: 0,
+            data: "Hello world",
+            children: ["test3"],
+            collapsed: true
+        }
+        items["test3"] = {
+            lastModifiedUnixMillis: 0,
+            data: "I am a child",
+            children: [],
+            collapsed: true
+        }
+        return items;
+    }
+}
+
+const getDataBlockWithVirtualRoot = (): BaseStoreDataType => {
+    return {
+        [virtualRootId]: {
+            lastModifiedUnixMillis: 0,
+            data: "",
+            children: [],
+            collapsed: false
+        }
     }
 }